Create a new GatewayInvalidParametersRefFeature and move GatewayInvalidParametersRef test to a new suite
