+++ pkg/features/gateway.go
@@ -62,9 +62,9 @@ const (
 	// HTTP 421 (Misdirected Request) when appropriate.
 	SupportGatewayHTTPSListenerDetectMisdirectedRequests = "GatewayHTTPSListenerDetectMisdirectedRequests"
 
-	// SupportGatewayInfrastructurePropagation option indicates support for
-	// spec.infrastructure.annotations and spec.infrastructure.labels
-	SupportGatewayInfrastructurePropagation FeatureName = "GatewayInfrastructurePropagation"
+	// SupportGatewayInfrastructure option indicates support for
+	// spec.infrastructure
+	SupportGatewayInfrastructure FeatureName = "GatewayInfrastructure"
 
 	// SupportGatewayAddressEmpty option indicates support for an empty
 	// spec.addresses.value field
@@ -110,9 +110,9 @@ var (
 		Name:    SupportGatewayHTTPSListenerDetectMisdirectedRequests,
 		Channel: FeatureChannelStandard,
 	}
-	// GatewayInfrastructurePropagationFeature contains metadata for the GatewayInfrastructurePropagation feature.
-	GatewayInfrastructurePropagationFeature = Feature{
-		Name:    SupportGatewayInfrastructurePropagation,
+	// GatewayInfrastructureFeature contains metadata for the GatewayInfrastructure feature.
+	GatewayInfrastructureFeature = Feature{
+		Name:    SupportGatewayInfrastructure,
 		Channel: FeatureChannelStandard,
 	}
 	// GatewayEmptyAddressFeature contains metadata for the SupportGatewayAddressEmpty feature.
@@ -152,7 +152,7 @@ var GatewayExtendedFeatures = sets.New(
 	GatewayStaticAddressesFeature,
 	GatewayHTTPListenerIsolationFeature,
 	GatewayHTTPSListenerDetectMisdirectedRequestsFeature,
-	GatewayInfrastructurePropagationFeature,
+	GatewayInfrastructureFeature,
 	GatewayEmptyAddressFeature,
 	GatewayBackendClientCertificateFeature,
 	GatewayFrontendClientCertificateValidationFeature,
